Update a specific cell with time stamp
This is useful in the case where you need to check if the a time trigger function is working or not. Just add this function after the scheduled function and this will update the time stamp in a particular cell. e.g i create and schedule a hourly trigger function   deleteDocByName(). To be sure that this runs, i created a function ExecSeq() and put the timestamp function Edit() inside the ExecSeq(). Then i scheduled this as hourly trigger.

function ExecSeq() {
  deleteDocByName();
  Edit()
}